The inclusion of graalvm fixes T6763
Summary:
Inital version, it build and the files get copied but need some advice on where to store the generated files, this one is a bit more complicated as it can replace other languages like java.
Could be a case of install it and leave it to the user to set ENV.
It also has quite high ram requirements to build.
fixes T6763.

Test Plan: Generated eopkg and installed, will need further testing

Reviewers: #triage_team, Girtablulu

Reviewed By: #triage_team, Girtablulu

Subscribers: Girtablulu, nazar, DataDrake, ReillyBrogan

Maniphest Tasks: T6763

Differential Revision: https://dev.getsol.us/D12495